+++ apps/docs/src/app/core/api-files.ts
@@ -353,6 +353,11 @@ export const API_FILES = {
     switch: ['SwitchComponent'],
     stepInput: ['StepInputComponent'],
     slider: ['SliderComponent'],
+    splitter: [
+        'SplitterComponent',
+        'SplitterPaneContainerComponent',
+        'SplitterSplitPaneComponent'
+    ],
     token: ['TokenComponent', 'TokenizerComponent'],
     toolbar: [
         'ToolbarComponent',