+++ e2e/wdio/core/tests/inline-help.e2e-spec.ts
@@ -41,7 +41,8 @@ describe('Inline help test suite', function() {
         expect(getAttributeByName(inlineHelpInput, 'fd-inline-help')).toContain('Inline Help Tooltip');
     });
 
-    it('Verify button inline help', () => {
+    // skip due to visual regression
+    xit('Verify button inline help', () => {
         scrollIntoView(exampleAreaContainersArr, 1);
         click(inlineHelpButton);
         waitForPresent('fd-popover-body');
@@ -52,7 +53,8 @@ describe('Inline help test suite', function() {
         expect(getAttributeByName(inlineHelpButton, 'fd-inline-help')).toContain('Inline Help Tooltip');
     });
 
-    it('Verify styled inline help icon', () => {
+    // skip due to visual regression
+    xit('Verify styled inline help icon', () => {
         scrollIntoView(exampleAreaContainersArr, 2);
         mouseHoverElement(inlineHelpStyledIcon);
         saveElementScreenshot(inlineHelpExampleExtended, `inline-help-styled-icon`, inlineHelpPage.getScreenshotFolder(), 2);
@@ -62,7 +64,8 @@ describe('Inline help test suite', function() {
         expect(getAttributeByName(inlineHelpStyledIcon, 'fd-inline-help')).toContain('Inline Help Tooltip');
     });
 
-    it('Verify template inline help example', () => {
+    // skip due to visual regression
+    xit('Verify template inline help example', () => {
         scrollIntoView(exampleAreaContainersArr, 3);
         mouseHoverElement(inlineHelpTemplateExample);
         saveElementScreenshot(inlineHelpExampleExtended, `inline-help-template`, inlineHelpPage.getScreenshotFolder(), 3);