+++ libs/core/src/lib/fundamental-ngx.module.ts
@@ -74,7 +74,7 @@ import { MenuModule, MenuMobileModule } from '@fundamental-ngx/core/menu';
 import { TabsModule } from '@fundamental-ngx/core/tabs';
 import { ResizableCardLayoutModule } from '@fundamental-ngx/core/resizable-card-layout';
 import { UploadCollectionModule } from '@fundamental-ngx/core/upload-collection';
-
+import { SplitterModule } from '@fundamental-ngx/core/splitter';
 import { PaginationModule } from '@fundamental-ngx/core/pagination';
 import { SelectMobileModule, SelectModule } from '@fundamental-ngx/core/select';
 import { ShellbarModule } from '@fundamental-ngx/core/shellbar';
@@ -86,6 +86,7 @@ import { MessagePageModule } from '@fundamental-ngx/core/message-page';
 import { VerticalNavigationModule } from '@fundamental-ngx/core/vertical-navigation';
 import { MicroProcessFlowModule } from '@fundamental-ngx/core/micro-process-flow';
 import { TimelineModule } from '@fundamental-ngx/core/timeline';
+
 
 @NgModule({
     imports: [CommonModule, FormsModule],
@@ -175,6 +176,7 @@ import { TimelineModule } from '@fundamental-ngx/core/timeline';
         MessageToastModule,
         DynamicSideContentModule,
         SliderModule,
+        SplitterModule,
         OnlyDigitsModule,
         IllustratedMessageModule,
         StatusIndicatorModule,